店舗推薦ロジックを別ファイルに切り出し
